split test files by group
