+++ tests/serving/openai/test_responses.py
@@ -12,18 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for set_openai_frontend() and OpenAI endpoint group body mappings."""
-
-from http import HTTPMethod
-from typing import cast
+"""Mock-server tests for the Responses endpoint group."""
 
 import pytest
 
 import mlrun
-from mlrun.runtimes.nuclio.serving import ServingRuntime
-from mlrun.serving.endpoint_mapping import APIHandlerConfig
-from mlrun.serving.openai_mappings import ENDPOINT_CLASSES, OpenAIEndpoint
-from mlrun.serving.server import GraphServer
+from mlrun.serving.openai_mappings import OpenAIEndpoint
 from tests.serving.assets.openai_fixtures import (
     COMPACT_EXPECTED_KWARGS,
     COMPACT_EXPECTED_RESPONSE,
@@ -42,93 +36,7 @@ from tests.serving.assets.openai_fixtures import (
     RESPONSE_OBJECT_EXPECTED_RESPONSE,
     RESPONSE_OBJECT_HANDLER_RESPONSE,
 )
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _make_fn() -> ServingRuntime:
-    return cast(ServingRuntime, mlrun.new_function("test-openai", kind="serving"))
-
-
-def _config(fn: ServingRuntime) -> APIHandlerConfig:
-    return APIHandlerConfig.from_dict(fn.spec.api_handler_config)
-
-
-def _make_mock_server(endpoint_group: OpenAIEndpoint, handler) -> GraphServer:
-    fn = _make_fn()
-    fn.set_openai_frontend([endpoint_group])
-    graph = fn.set_topology("flow", engine="sync")
-    graph.to(name="handler", handler=handler).respond()
-    return fn.to_mock_server()
-
-
-# ---------------------------------------------------------------------------
-# Registry structure
-# ---------------------------------------------------------------------------
-class TestOpenAIRegistry:
-    def test_all_groups_present(self) -> None:
-        """Registry contains an entry for every OpenAIEndpoint value."""
-        for group in OpenAIEndpoint:
-            assert group in ENDPOINT_CLASSES
-
-
-# ---------------------------------------------------------------------------
-# set_openai_frontend() wiring
-# ---------------------------------------------------------------------------
-class TestSetOpenAIFrontend:
-    def test_responses_only_registers_correct_endpoints(self) -> None:
-        """set_openai_frontend([RESPONSES]) registers exactly the Responses endpoints."""
-        fn = _make_fn()
-        fn.set_openai_frontend([OpenAIEndpoint.RESPONSES])
-
-        config = _config(fn)
-        responses_endpoints = ENDPOINT_CLASSES[OpenAIEndpoint.RESPONSES].endpoints()
-
-        for ep in responses_endpoints:
-            endpoint = config.get_endpoint_config(ep["http_method"], ep["path"])
-            assert endpoint is not None, (
-                f"Expected {ep['http_method']} {ep['path']} to be registered"
-            )
-
-        assert len(config.endpoints) == len(responses_endpoints)
-
-    def test_default_registers_all_groups(self) -> None:
-        """set_openai_frontend() with no args registers all OpenAIEndpoint groups."""
-        fn = _make_fn()
-        fn.set_openai_frontend()
-
-        config = _config(fn)
-        for group in OpenAIEndpoint:
-            for ep in ENDPOINT_CLASSES[group].endpoints():
-                endpoint = config.get_endpoint_config(ep["http_method"], ep["path"])
-                assert endpoint is not None, (
-                    f"Expected {ep['http_method']} {ep['path']} to be registered"
-                )
-
-    def test_preserves_existing_config(self) -> None:
-        """set_openai_frontend() merges into an existing APIHandlerConfig."""
-        from mlrun.common.schemas.serving import APIHandlerAction
-
-        existing = APIHandlerConfig()
-        existing.add_endpoint_handler("/health", HTTPMethod.GET, APIHandlerAction.ALLOW)
-        fn = _make_fn()
-        fn.set_api_handler_config(existing)
-
-        fn.set_openai_frontend([OpenAIEndpoint.RESPONSES])
-
-        config = _config(fn)
-        assert config.get_endpoint_config(HTTPMethod.GET, "/health") is not None
-        assert (
-            config.get_endpoint_config(HTTPMethod.POST, "/responses/compact")
-            is not None
-        )
-
-
-# ---------------------------------------------------------------------------
-# Responses group — mock server tests
-# ---------------------------------------------------------------------------
+from tests.serving.openai.openai_common import make_mock_server
 
 
 class TestResponsesGroupMock:
@@ -147,7 +55,7 @@ class TestResponsesGroupMock:
             captured.update(kwargs)
             return RESPONSE_OBJECT_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test("/responses", method="POST", body=CREATE_REQUEST_BODY)
             assert "extra_field" not in captured
@@ -165,7 +73,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -186,7 +94,7 @@ class TestResponsesGroupMock:
             captured_kwargs.update(kwargs)
             return RESPONSE_OBJECT_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test(f"/responses/{RESPONSE_ID}", method="GET")
             assert captured_kwargs.get("response_id") == RESPONSE_ID
@@ -202,7 +110,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -221,16 +129,15 @@ class TestResponsesGroupMock:
 
         def handler(body, **kwargs):
             captured_kwargs.update(kwargs)
-            return {**DELETE_HANDLER_RESPONSE, "extra_field": "should_be_filtered"}
+            return DELETE_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test(f"/responses/{RESPONSE_ID}", method="DELETE")
             assert captured_kwargs.get("response_id") == RESPONSE_ID
             assert resp["id"] == RESPONSE_ID
             assert resp["deleted"] is True
             assert resp["object"] == "response"
-            assert "extra_field" not in resp
         finally:
             server.wait_for_completion()
 
@@ -246,9 +153,11 @@ class TestResponsesGroupMock:
             captured_kwargs.update(kwargs)
             return INPUT_ITEMS_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
-            resp = server.test(f"/responses/{RESPONSE_ID}/input_items", method="GET")
+            resp = server.test(
+                f"/responses/{RESPONSE_ID}/input_items", method="GET"
+            )
             assert captured_kwargs.get("response_id") == RESPONSE_ID
             assert "extra_field" not in resp
             for key, value in INPUT_ITEMS_EXPECTED_RESPONSE.items():
@@ -262,7 +171,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -284,7 +193,7 @@ class TestResponsesGroupMock:
             captured.update(kwargs)
             return INPUT_TOKENS_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test(
                 "/responses/input_tokens",
@@ -306,7 +215,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -327,7 +236,7 @@ class TestResponsesGroupMock:
             captured_kwargs.update(kwargs)
             return RESPONSE_OBJECT_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test(f"/responses/{RESPONSE_ID}/cancel", method="POST")
             assert captured_kwargs.get("response_id") == RESPONSE_ID
@@ -343,7 +252,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -365,7 +274,7 @@ class TestResponsesGroupMock:
             captured.update(kwargs)
             return COMPACT_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             resp = server.test(
                 "/responses/compact", method="POST", body=COMPACT_REQUEST_BODY
@@ -385,7 +294,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return {}
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(
                 mlrun.errors.MLRunBadRequestError, match="Mandatory field"
@@ -404,7 +313,7 @@ class TestResponsesGroupMock:
         def handler(body, **kwargs):
             return COMPACT_HANDLER_RESPONSE
 
-        server = _make_mock_server(OpenAIEndpoint.RESPONSES, handler)
+        server = make_mock_server(OpenAIEndpoint.RESPONSES, handler)
         try:
             with pytest.raises(RuntimeError, match="Mandatory field"):
                 server.test(
@@ -414,42 +323,3 @@ class TestResponsesGroupMock:
                 )
         finally:
             server.wait_for_completion()
-
-
-# ---------------------------------------------------------------------------
-# ChatCompletions group — placeholder (ML-12461)
-# ---------------------------------------------------------------------------
-class TestChatCompletionsGroupMock:
-    """Mock tests for the ChatCompletions endpoint group (endpoints TBD — ML-12461)."""
-
-    def test_no_endpoints_registered(self) -> None:
-        """ChatCompletions group has no endpoints yet."""
-        fn = _make_fn()
-        fn.set_openai_frontend([OpenAIEndpoint.CHAT_COMPLETIONS])
-        assert len(_config(fn).endpoints) == 0
-
-
-# ---------------------------------------------------------------------------
-# Audio group — placeholder (ML-12461)
-# ---------------------------------------------------------------------------
-class TestAudioGroupMock:
-    """Mock tests for the Audio endpoint group (endpoints TBD — ML-12461)."""
-
-    def test_no_endpoints_registered(self) -> None:
-        """Audio group has no endpoints yet."""
-        fn = _make_fn()
-        fn.set_openai_frontend([OpenAIEndpoint.AUDIO])
-        assert len(_config(fn).endpoints) == 0
-
-
-# ---------------------------------------------------------------------------
-# Images group — placeholder (ML-12461)
-# ---------------------------------------------------------------------------
-class TestImagesGroupMock:
-    """Mock tests for the Images endpoint group (endpoints TBD — ML-12461)."""
-
-    def test_no_endpoints_registered(self) -> None:
-        """Images group has no endpoints yet."""
-        fn = _make_fn()
-        fn.set_openai_frontend([OpenAIEndpoint.IMAGES])
-        assert len(_config(fn).endpoints) == 0
